[FM] Temporary: use software rendering for FM-to-earphone playback
The Qualcomm audio HAL does not yet support volume control when
direct FM-to-sink patches are created. Force use of software rendering
(reading from AudioRecord and writing to AudioTrack) for now, as this
has working software volume control.

This change is obviously bad for power, as it requires SW involvement
in the audio path. The change should be reverted when source HW volume
control is available with audio patches.

Playback through speaker, BT, and other sinks already go use software
rendering and should be unaffected by this change.

Change-Id: Ib732c8e8328035d2b3e60b7823446d1d75b3c107